% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{fhl}
\alias{fhl}
\title{Farre-Mensa, Hegde, and Ljungqvist (2020) data}
\format{
A data frame with 34,435 observations on 11 variables:

\describe{

\item{gau}{Group Art Unit ID of the worker assigned to application}

\item{examiner}{ID of worker assigned to application}

\item{lenience}{Leniency measure: number of patents allowed / number of
                patents examined to date.}

\item{pat_appl_year}{Patent application year}

\item{dallowed}{First patent application approved}

\item{ln_total_patents_appl}{log(1 + subsequent patent applications)}

\item{state_num}{State code}

\item{ln_total_patents_approved}{log(1 + subsequent patent approvals)}

\item{ln_all_cites5}{log(1 + total citations to subsequent patent
                     applications)}

\item{ln_rounds_before_patent_APP}{log(1 + number of prior VC rounds)}

\item{ind_year}{Art-unit-by-application-year fixed effects.}

}
}
\source{
Journal of Finance website, \doi{10.1111/jofi.12867}
}
\usage{
fhl
}
\description{
TODO: describe
}
\references{
{

\cite{Farre-Mensa, Hegde, and Ljungqvist. 2020. What Is a Patent Worth?
Evidence from the U.S. Patent "Lottery". The Journal of Finance 72
(2):639-682. \doi{10.1111/jofi.12867}.}

}
}
\keyword{datasets}
