% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVData}
\alias{IVData}
\title{Class constructor for IVData}
\usage{
IVData(Y, Z, W)
}
\arguments{
\item{Y}{[n x 2] class \code{Matrix}}

\item{Z}{[n x k] Matrix of instruments}

\item{W}{[n x ell] Matrix of covariates}
}
\description{
Convert data to standardized format for use with low-level functions. Uses
\code{Matrix} package, which speeds up calculations.
}
