% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVregIL.fit}
\alias{IVregIL.fit}
\title{Inference based on invariant likelihood}
\usage{
IVregIL.fit(d)
}
\arguments{
\item{d}{object of class \code{RDData}}
}
\description{
Inference based on invariant likelihood
}
