% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{IVregLI.fit}
\alias{IVregLI.fit}
\title{Limited information likelihood}
\usage{
IVregLI.fit(d)
}
\arguments{
\item{d}{object of class \code{RDData}}
}
\description{
Limited information likelihood
}
\keyword{internal}
