% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{IVregRE.fit}
\alias{IVregRE.fit}
\title{Random-effects likelihood}
\usage{
IVregRE.fit(d)
}
\arguments{
\item{d}{object of class \code{RDData}}
}
\description{
Random-effects likelihood
}
\keyword{internal}
