% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ak80}
\alias{ak80}
\title{Angrist-Krueger (1991) census data}
\format{
A data frame with 329,509 observations on 10 variables:

\describe{

\item{age}{Age, measured at quarterly precision}

\item{education}{Years of education}

\item{lwage}{Log of weekly wage}

\item{married}{Indicator for being married}

\item{qob}{Quarter of birth}

\item{sob}{State of birth (FIPS code)}

\item{black}{Indicator for being black}

\item{smsa}{SMSA indicator}

\item{yob}{Year of birth}

\item{division}{Factor variable indicating division}

}
}
\source{
Josh Angrist's website,
    \url{http://economics.mit.edu/faculty/angrist/data1/data/angkru1991}
}
\usage{
ak80
}
\description{
Sample of males born in the US in 1930-39 from 5% sample of the 1980 Census
}
\keyword{datasets}
