% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVData}
\alias{IVData}
\title{Class constructor for \code{IVData}}
\usage{
IVData(Y, X, Z, W, moments = TRUE, approx = TRUE)
}
\arguments{
\item{Y}{n-vector}

\item{X}{n-vector}

\item{Z}{[n x k] Matrix of instruments, class \code{Matrix}}

\item{W}{[n x ell] Matrix of covariates, class \code{Matrix}}

\item{moments}{if \code{TRUE}, compute estimates of third and fourth moments
of the reduced-form errors based on least squares residuals}

\item{approx}{if \code{TRUE}, then estimates of third and fourth moments use
an approximation to speed up the calculations.}
}
\description{
Convert data to standardized format for use with low-level functions. Uses
\code{Matrix} package, which speeds up calculations.
}
