% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVoverid}
\alias{IVoverid}
\title{Test of overidentifying restrictions}
\usage{
IVoverid(r)
}
\arguments{
\item{r}{object of class \code{RDResults}}
}
\description{
Report test statistic and p-value for testing of overidentifying
restrictions. The Sargan test is valid under few instruments. The Modified
Cragg-Donald test (Modified-CD) corresponds to a test due to Cragg and Donald
(1993), with a critical value modified to make it robust to many instruments
and many exogenous regressors.
}
