% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVreg.fit}
\alias{IVreg.fit}
\title{Low-level interfact to \code{IVreg}}
\usage{
IVreg.fit(d, inference)
}
\arguments{
\item{d}{Object of class \code{"IVData"}}

\item{inference}{Vector specifying inference method(s). The elements of
    the vector can consist of the following methods:
\describe{

  \item{"standard"}{Report inference based on tsls, liml, and mbtsls, along
                    with homoscedastic heteroscedasticity-robust standard
                    errors valid under standard asymptotic sequence}

  \item{"re"}{Inference based on Hessian of random effects likelihood}

  \item{"il"}{Inference based on Hessian of invariant likelihood, evaluated
              numerically}

  \item{"lil"}{Inference based on information matrix of limited
              information likelihood}

  \item{"md"}{Inference based on the minimum distance objective function} }}
}
\description{
Low-level interfact to \code{IVreg}
}
