% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyiv.R
\name{IVreg}
\alias{IVreg}
\title{Fit instrumental-variable regression}
\usage{
IVreg(formula, data, subset, na.action, inference = "standard",
  approx = TRUE)
}
\arguments{
\item{formula}{specification of the regression relationship and the
instruments of the form \code{y ~ x + w1 + w2 | z1 + z2 + z3}, where
\code{y} is the outcome variable, \code{x} is a scalar endogenous
variable, \code{w1,w2} are exogenous regressors, and \code{z1,z2,z3} are
excluded instruments.}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{inference}{Vector specifying inference method(s). The elements of
    the vector can consist of the following methods:
\describe{

  \item{"standard"}{Report inference based on tsls, liml, and mbtsls, along
                    with homoscedastic heteroscedasticity-robust standard
                    errors valid under standard asymptotic sequence}

  \item{"re"}{Inference based on Hessian of random effects likelihood}

  \item{"il"}{Inference based on Hessian of invariant likelihood, evaluated
              numerically}

  \item{"lil"}{Inference based on information matrix of limited
              information likelihood}

  \item{"md"}{Inference based on the minimum distance objective function} }}

\item{approx}{if \code{TRUE}, then estimates of third and fourth moments used
in inference based on the minimum distance objective function
(\code{inference="md"}) use an appriximation to speed up the
calculations.}
}
\description{
Fit instrumental-variable regression
}
\examples{
## Specification as in Table V, columns (1) and (2) in Angrist and Krueger
IVreg(lwage~education+as.factor(yob)|as.factor(qob)*as.factor(yob),
           data=ak80, inference=c("standard", "re", "il", "lil"))
## Only quarter of birth as instrument, add married, black and smsa as exogenous
#regressors
IVreg(lwage~education+as.factor(yob)+black+smsa+married|as.factor(qob),
           data=ak80, inference=c("standard", "re", "il", "lil"))
}
